Updated to v3.2.2

+++ src/main/java/com/github/hhhzzzsss/songplayer/song/NBSConverter.java
@@ -31,7 +31,7 @@ public class NBSConverter {
         public byte instrument;
         public byte key;
         public byte velocity = 100;
-        public int panning = 100;
+        public byte panning = 100;
         public short pitch = 0;
     }
 
@@ -105,7 +105,7 @@ public class NBSConverter {
                 note.key = buffer.get();
                 if (format >= 4) {
                     note.velocity = buffer.get();
-                    note.panning = buffer.get() & 0xff;
+                    note.panning = (byte) (buffer.get() & 0xff);
                     note.pitch = buffer.getShort();
                 }
                 nbsNotes.add(note);
